fix: fix script
