+++ apps/desktop/electron-builder-base.config.js
@@ -1,0 +1,58 @@
+const path = require('path');
+
+/* eslint-disable no-template-curly-in-string */
+require('../../development/env');
+
+const getPath = (...paths) => {
+  console.log(path.join(__dirname, ...paths));
+  return path.join(__dirname, ...paths);
+};
+
+const baseElectronBuilderConfig = {
+  'extraMetadata': {
+    'main': 'dist/app.js',
+    'version': process.env.VERSION,
+  },
+  'appId': 'so.onekey.wallet.desktop',
+  'productName': 'OneKey',
+  'copyright': 'Copyright © ${author}',
+  'asar': true,
+  'buildVersion': process.env.BUILD_NUMBER,
+  'directories': {
+    'output': 'build-electron',
+  },
+  'npmRebuild': false,
+  'files': [
+    'dist/**/*',
+    '!dist/__**',
+    'build/**/*',
+    '!build/static/bin/**/*',
+    'package.json',
+  ],
+  'protocols': {
+    'name': 'electron-deep-linking',
+    'schemes': ['onekey-wallet', 'wc', 'ethereum'],
+  },
+  'extraResources': [
+    {
+      'from': 'app/build/static/images/icons/512x512.png',
+      'to': 'static/images/icons/512x512.png',
+    },
+    {
+      'from': 'build/static/images/icons/icon.icns',
+      'to': 'static/images/icons/icon.icns',
+    },
+    {
+      'from': 'app/build/static/preload.js',
+      'to': 'static/preload.js',
+    },
+  ],
+  'publish': {
+    'provider': 'github',
+    'repo': 'app-monorepo',
+    'owner': 'OneKeyHQ',
+  },
+  'afterSign': getPath('scripts/notarize.js'),
+  'afterPack': getPath('scripts/fileOperation.js'),
+};
+module.exports = baseElectronBuilderConfig;
