+++ apps/desktop/electron-builder-snap.config.js
@@ -1,56 +1,17 @@
 /* eslint-disable no-template-curly-in-string */
 require('../../development/env');
+const baseElectronBuilderConfig = require('./electron-builder-base.config');
 
 module.exports = {
-  'extraMetadata': {
-    'main': 'dist/app.js',
-    'version': process.env.VERSION,
-  },
-  'appId': 'so.onekey.wallet.desktop',
-  'productName': 'OneKey',
-  'copyright': 'Copyright © ${author}',
-  'asar': true,
-  'buildVersion': process.env.BUILD_NUMBER,
-  'directories': {
-    'output': 'build-electron',
-  },
-  'files': [
-    'build/**/*',
-    '!build/static/bin/**/*',
-    'dist/**/*.js',
-    '!dist/__**',
-    'package.json',
-  ],
-  'protocols': {
-    'name': 'electron-deep-linking',
-    'schemes': ['onekey-wallet', 'wc', 'ethereum'],
-  },
-  'extraResources': [
-    {
-      'from': 'build/static/images/icons/512x512.png',
-      'to': 'static/images/icons/512x512.png',
-    },
-    {
-      'from': 'build/static/preload.js',
-      'to': 'static/preload.js',
-    },
-  ],
-  'publish': {
-    'provider': 'github',
-    'repo': 'app-monorepo',
-    'owner': 'OneKeyHQ',
-  },
-  'snap': {
-    'grade': 'stable',
-  },
+  ...baseElectronBuilderConfig,
   'linux': {
     'extraResources': [
       {
-        'from': 'build/static/bin/bridge/linux-${arch}',
+        'from': 'app/build/static/bin/bridge/linux-${arch}',
         'to': 'bin/bridge',
       },
     ],
-    'icon': 'build/static/images/icons/512x512.png',
+    'icon': 'app/build/static/images/icons/512x512.png',
     'artifactName': 'OneKey-Wallet-${version}-linux-${arch}.${ext}',
     'executableName': 'onekey-wallet',
     'category': 'Utility',
