+++ packages/components/src/layouts/Navigation/Navigator/ModalFlowNavigator.tsx
@@ -69,7 +69,6 @@ function ModalFlowNavigator<RouteName extends string, P extends ParamListBase>({
   }, [onMounted, onUnmounted]);
 
   return (
-    // @ts-expect-error
     <ModalStack.Navigator screenOptions={makeScreenOptions}>
       {config.map(
         ({
